gpllib: use four bytes to discover the MBR
Syslinux and Master Boot LoaDeR share the first 2 bytes. Use one more
to distinguish between the two of them.

To keep standards types, uint16_t is being replaced by uint32_t.
The extra byte is not used yet.

+++ com32/gplinclude/disk/bootloaders.h
@@ -13,6 +13,6 @@
 #include <stdint.h>
 #include <disk/geom.h>
 
-void get_bootloader_string(const uint16_t, void *, const int);
-uint16_t get_bootloader_id(const struct driveinfo *);
+void get_bootloader_string(const uint32_t, void *, const int);
+uint32_t get_bootloader_id(const struct driveinfo *);
 #endif /* _BOOTLOADERS_H_ */
